refactor: update angular.json & co
