+++ public/content/images/angular-christmas-calendar.png
