+++ public/content/images/angular-space-footer.png
