+++ public/content/images/angular-space.png
