+++ public/content/images/angular_wordmark_gradient.png
