+++ src/app/feature/calendar/calendar-grid.component.html
@@ -1,6 +1,6 @@
 <div class="title">
   <img
-    ngSrc="./../../../assets/content/images/angular_wordmark_gradient.png"
+    ngSrc="content/images/angular_wordmark_gradient.png"
     width="400"
     height="125"
     priority="true"
@@ -67,7 +67,7 @@
 <div class="card space">
   <a href="https://angularspace.com/?ref=angularchristmascalendar.com" target="_blank">
     <img
-      ngSrc="./../../../assets/content/images/angular-space.png"
+      ngSrc="content/images/angular-space.png"
       width="100"
       height="100"
       href="https://polarized-pram-52c.notion.site/What-s-the-First-Name-of-Angular-Veteran-who-has-become-AI-Expert-150d2d2ccbe0800bb4d1e8aab045f89c?pvs=4"
