+++ src/index.html
@@ -22,7 +22,7 @@
     />
     <meta
       property="og:image"
-      content="https://angularchristmascalendar.com/assets/content/images/angular-christmas-calendar.png"
+      content="https://angularchristmascalendar.com/content/images/angular-christmas-calendar.png"
     />
 
     <!-- Twitter -->
@@ -35,7 +35,7 @@
     />
     <meta
       property="twitter:image"
-      content="https://angularchristmascalendar.com/assets/content/images/angular-christmas-calendar.png"
+      content="https://angularchristmascalendar.com/content/images/angular-christmas-calendar.png"
     />
 
     <link rel="icon" type="image/x-icon" href="favicon.ico" />
@@ -58,7 +58,7 @@
         <p style="margin-bottom: 4px">
           <a href="https://angularspace.com/?ref=angularchristmascalendar.com" target="_blank">
           <img
-            src="./assets/content/images/angular-bros.png"
+            src="content/images/angular-bros.png"
             alt="Angular Bros"
             style="margin-right: 15px; opacity: 0.8"
             width="132"
