+++ src/styles.css
@@ -35,7 +35,7 @@
 }
 
 body {
-  background: url('./assets/content/images/background.jpg') no-repeat top center var(--white);
+  background: url('assets/content/images/background.jpg') no-repeat top center var(--white);
   background-size: cover;
   overflow-x: hidden;
 }
