+++ tsconfig.app.json
@@ -11,7 +11,7 @@
   "files": [
     "src/main.ts",
     "src/main.server.ts",
-    "server.ts"
+    "src/server.ts"
   ],
   "include": [
     "src/**/*.d.ts"
